requirements.txt
# requirements.txt

allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
